types: Moved ButtonKeyCode, MouseButton, KeyCode, MouseWheelDelta and PlayerEvent from core to types
