+++ types/src/lib.rs
@@ -1,5 +1,9 @@
+#[macro_use]
+extern crate num_derive;
+
 pub mod backend;
 pub mod binary_data;
+pub mod events;
 pub mod numbers;
 pub mod string;
 pub mod tag_utils;
